ImageMagick: only build the jpeg coder.
The sole user of ImageMagick (the legacy wallpapers) only needs jpeg, so
let's get out of the business of supporting every other format known to
man.

Test: wallpaper app still works on taimen
Change-Id: I7c5673a90aec445c6e03c3557aec4dc780f8df0a

ImageMagick: have DDS and XC coders added

The user of ImageMagick (the legacy wallpapers) needs DDS and XC coders.

Also move the Android-specific list of coders to the top of
coders-list.h to avoid merge conflicts.

Test: wallpaper should run in walleye
Change-Id: If8915d51ac2f3e7a4427241e5dfbe52a2f5b5447


Bug: 136733674